backend/requirements.txt
flask==2.3.3
python-dotenv==0.19.0
requests==2.31.0
beautifulsoup4==4.12.3
supabase==1.0.3
APScheduler==3.10.4
pytest==8.3.4
pytest-asyncio==0.23.8
pytest-mockito==0.0.4
python-dateutil==2.9.0
fake-http-header==0.3.5
resend==0.7.2

